# server/openapi.yaml
openapi: 3.1.0
info:
  title: Trier OS API
  version: 3.5.1
  description: |
    Enterprise Industrial Operating System REST API.
    Single-instance deployment — all plants connect to one corporate server.
    Plant scope is set per-request via the x-plant-id header.
  contact:
    name: Trier OS Engineering
  license:
    name: Proprietary — © 2026 Trier OS. All Rights Reserved.
servers:
  - url: /api
    description: Trier OS corporate server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: query
      name: api_key
  parameters:
    PlantIdHeader:
      name: x-plant-id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]{1,64}$'
      description: Plant database identifier. Use 'all_sites' for corporate-scoped endpoints.
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /loto/permits:
    get:
      summary: "List LOTO permits (supports ?status=open&assetId=)"
      operationId: getlotoPermits
      tags:
        - Safety & Compliance
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
    post:
      summary: "Open a new LOTO permit"
      operationId: postlotoPermits
      tags:
        - Safety & Compliance
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /loto/permits/{id}:
    get:
      summary: "Get single LOTO permit with isolation points"
      operationId: getlotoPermitsById
      tags:
        - Safety & Compliance
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
    put:
      summary: "Update or close a LOTO permit"
      operationId: putlotoPermitsById
      tags:
        - Safety & Compliance
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
...